backend/requirements.txt
asgiref==3.6.0
Django==3.1.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.7
dnspython==2.7.0
pip==21.2.3
PyJWT==2.10.1
pymongo==3.11.4
pytz==2025.1
setuptools==57.4.0
sqlparse==0.2.4
tzdata==2025.1
python-dotenv